% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawMS.R
\name{drawMS}
\alias{drawMS}
\title{drawMS}
\usage{
drawMS(
  rawDataName,
  mzRange = c(700, 850),
  rtRange = c(0, Inf),
  mzTol = 5e-04,
  msLevel = 1,
  mzLabelDis = 20,
  intLabelTol = 3,
  res = 600,
  dcm = 2
)
}
\arguments{
\item{rawDataName}{character. File name of the raw data (mzML or mzXML).}

\item{mzRange}{vector. m/z range to plot.}

\item{rtRange}{vector. Retention time to average.}

\item{mzTol}{float. m/z tolerance to average spectra.}

\item{msLevel}{int. MS level (1, 2 or 3).}

\item{mzLabelDis}{float. A value to specify the range that only one m/z will be labeled, in Da.}

\item{intLabelTol}{float. Relative intensity larger than this value can be labeled.}

\item{res}{int. Resolution of the output figure.}

\item{dcm}{int. Number of decimals for m/z labels.}
}
\value{
This function outputs a figure in the current working directory.
}
\description{
Draw an averaged MS scan during a certain retention time.
}
\examples{
# drawMS(rawDataName)
}
