% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIntensity.R
\name{getIntensity}
\alias{getIntensity}
\title{getIntensity}
\usage{
getIntensity(
  rawDataName,
  ionTableName,
  mzTol = 0.005,
  msLevel = 1,
  output = TRUE
)
}
\arguments{
\item{rawDataName}{character. File name of raw MS data (mzML or mzXML).}

\item{ionTableName}{character. A csv file specifying the m/z values and retention time windows.}

\item{mzTol}{float. m/z tolerance to average spectra.}

\item{msLevel}{int. MS level (1 or 2).}

\item{output}{Boolean. Set to \code{TRUE} then output the result table.}
}
\value{
This function either returns the result table or output it to csv file.
}
\description{
Extract intensities by m/z value and retention time
}
\examples{
# getInensity(rawDataName, ionTableName)
}
